+++ src/flask_sqlalchemy/__init__.py
@@ -1,9 +1,26 @@
 from __future__ import annotations
 
-from .extension import SQLAlchemy
+import typing as t
 
-__version__ = "3.1.0"
+from .extension import SQLAlchemy
 
 __all__ = [
     "SQLAlchemy",
 ]
+
+
+def __getattr__(name: str) -> t.Any:
+    if name == "__version__":
+        import importlib.metadata
+        import warnings
+
+        warnings.warn(
+            "The '__version__' attribute is deprecated and will be removed in"
+            " Flask-SQLAlchemy 3.2. Use feature detection or"
+            " 'importlib.metadata.version(\"flask-sqlalchemy\")' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return importlib.metadata.version("flask-sqlalchemy")
+
+    raise AttributeError(name)